requirements.txt
# ══════════════════════════════════════════════════════
#  AI Interview Simulator — Requirements
#  Python 3.12+ compatible (tested on 3.12 & 3.13)
# ══════════════════════════════════════════════════════

# ── Web Framework (Backend) ───────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# ── Database ──────────────────────────────────────────
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# ── AI / LLM ──────────────────────────────────────────
groq==0.13.0

# ── Resume Parsing ────────────────────────────────────
pdfplumber==0.11.4
python-docx==1.1.2

# ── Frontend ──────────────────────────────────────────
streamlit==1.41.0
plotly==5.24.1

# ── HTTP Client ───────────────────────────────────────
requests==2.32.3
httpx==0.27.2

# ── Data / Utils ──────────────────────────────────────
pandas==2.2.3
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# ── Dev / Optional ────────────────────────────────────
# Uncomment for local development:
# pytest==8.3.3
# pytest-asyncio==0.24.0
# black==24.10.0
# ruff==0.8.0
